stac-2931 added different tests for agent and receiver
