+++ test/integration/tests/agent_integration/molecule/default/tests/test_agent.py
@@ -1,0 +1,14 @@
+import os
+import pytest
+import testinfra.utils.ansible_runner
+
+
+testinfra_hosts = testinfra.utils.ansible_runner.AnsibleRunner(
+    os.environ['MOLECULE_INVENTORY_FILE']).get_hosts('agent_vm')
+
+
+
+def test_opt_stackstate_directory(host):
+    f = host.file('/opt/datadog-agent/')
+    
+    assert f.is_directory
