+++ test/integration/tests/agent_integration/molecule/default/tests/test_receiver.py
@@ -1,0 +1,18 @@
+import os
+import pytest
+import testinfra.utils.ansible_runner
+
+testinfra_hosts = testinfra.utils.ansible_runner.AnsibleRunner(
+    os.environ['MOLECULE_INVENTORY_FILE']).get_hosts('receiver_vm')
+
+
+def test_etc_docker_directory(host):
+    f = host.file('/etc/docker/')
+
+    assert f.is_directory
+
+
+def test_docker_compose_file(host):
+    f = host.file('/home/ubuntu/docker-compose.yml')
+
+    assert f.is_file
